move training model to DeepJet_models.py
