+++ Train/DeepFlavourRegressionBinned.py
@@ -1,5 +1,15 @@
 from __future__ import absolute_import
 from __future__ import division
+from pdb import set_trace
+from argparse import ArgumentParser
+# argument parsing and bookkeeping
+parser = ArgumentParser('Run the training')
+parser.add_argument('inputDir')
+parser.add_argument('outputDir')
+parser.add_argument('-f','--force', action='store_true', help='overwrite the directory if there')
+parser.add_argument('--warm', help='pre-trained model')
+args = parser.parse_args()
+
 
 import matplotlib
 #if no X11 use below
@@ -15,19 +25,14 @@ from keras.optimizers import SGD
 ## to call it from cammand lines
 import sys
 import os
-from argparse import ArgumentParser
 import shutil
-from pdb import set_trace
 
-# argument parsing and bookkeeping
-parser = ArgumentParser('Run the training')
-parser.add_argument('inputDataCollection')
-parser.add_argument('outputDir')
-parser.add_argument('-f','--force', action='store_true', help='overwrite the directory if there')
-args = parser.parse_args()
+import sys
+print 'got command: %s' % ' '.join(sys.argv)
 
-inputData = os.path.abspath(args.inputDataCollection)
-outputDir=args.outputDir
+from glob import glob
+inputData = glob('%s/conversion.*.dc' % os.path.abspath(args.inputDir))
+outputDir = args.outputDir
 # create output dir
 
 if os.path.isdir(outputDir) and not args.force:
@@ -50,8 +55,8 @@ shutil.copyfile('../modules/DeepJet_models.py',outputDir+'DeepJet_models.py')
 # configure the in/out/split etc
 config_args = { #we might want to move it to an external file
    'testrun'   : False,
-   'nepochs'   : 1,
-   'batchsize' : 500,
+   'nepochs'   : 100,
+   'batchsize' : 2000,
    'startlearnrate' : 0.0005,
    'useweights' : False,
    'splittrainandtest' : 0.8,
@@ -78,9 +83,7 @@ callbacks = DeepJet_callbacks(
 from DataCollection import DataCollection
 from TrainData_deepCSV_PF_binned import TrainData_deepCSV_PF_Binned
 
-traind = DataCollection()
-traind.readFromFile(inputData)
-traind.setBatchSize(config_args['batchsize'])
+traind = sum(DataCollection(i) for i in inputData)
 traind.useweights=config_args['useweights']
 
 if config_args['testrun']:
@@ -101,111 +104,14 @@ inputs = (
 
 ####################################################
 #                                                  #
-#  To be moved to a function in DeepJet_models.py  #
+#                    MODEL                         #
 #                                                  #
 ####################################################
 
-from keras.layers import Dense, Dropout, Flatten, Convolution2D, concatenate, \
-   Convolution1D, Convolution3D, Reshape, LocallyConnected2D
-from keras.models import Model
-
-#unpack inputs
-glob, charged, neutral, svs, bin_global = inputs
-
-def make_layers(layer_type, args_list, dropout, 
-                layer_in, dropout_at_first=False, **kwargs):
-    ret = layer_in
-    first = not dropout_at_first
-    for args in args_list:
-        if first:
-            first = False
-        else:
-            ret = Dropout(dropout)(ret)
-        ret = layer_type(*args, **kwargs)(ret)
-    return ret    
-
-#build single bin convolutions
-kwargs  = {'kernel_initializer' : 'lecun_uniform',  'activation' : 'relu'}
-k = (1,1,1) #kernel
-charged = make_layers(Convolution3D, [[64, k], [32, k], [32, k], [8, k]], config_args['conv_dropout'], charged, **kwargs)
-svs     = make_layers(Convolution3D, [[64, k], [32, k], [32, k], [8, k]], config_args['conv_dropout'], svs, **kwargs)
-neutral = make_layers(Convolution3D, [[32, k], [16, k], [4 , k]], config_args['conv_dropout'], neutral, **kwargs)
-
-#flatten the single bins
-charged = Reshape((
-      int(charged.shape[1]), # [0] is the number of batches, set to None, 1 is the # of x bins, returns Dimension(5), we need to cast to int
-      int(charged.shape[2]), # 2 is the # of y bins
-      int(charged.shape[3]*charged.shape[4]),
-      ))(charged)
-neutral = Reshape((
-      int(neutral.shape[1]), 
-      int(neutral.shape[2]), 
-      int(neutral.shape[3]*neutral.shape[4]),
-      ))(neutral)
-svs = Reshape((
-      int(svs.shape[1]), 
-      int(svs.shape[2]), 
-      int(svs.shape[3]*svs.shape[4]),
-      ))(svs)
-
-for ib in [1,2]:
-   if not (charged.shape[ib] == neutral.shape[ib] == svs.shape[ib]):
-      raise ValueError('The number of bins along the axis %d should be consistent for charged, neutral and sv features' % ib)
-
-#merge the info from different sources, but from the same bin, into a single place
-binned_info = concatenate([charged,neutral,svs,bin_global]) #shape (?, 10, 10, 76)
-
-#shrink the info size
-nentries_per_bin = int(binned_info.shape[-1])
-k = (1,1)
-binned_info = make_layers(
-    Convolution2D, [[nentries_per_bin, k], [25, k]], 
-    config_args['conv_dropout'], binned_info, 
-    dropout_at_first=True, **kwargs
-    )
-
-#learn a different representation for each bin
-#and reduce the dimensionality to 3x3
-xbins, ybins = int(binned_info.shape[1]), int(binned_info.shape[2])
-kernel = xbins//3
-if xbins != ybins:
-    raise ValueError('The number of x and y bins should be the same!')
-if xbins % 3 != 0:
-    raise ValueError('The number of x and y bins should be a multiplier of 3')
-
-binned_info = make_layers(
-    LocallyConnected2D, [[40, kernel]], 
-    config_args['conv_dropout'], binned_info,
-    dropout_at_first=True, strides=kernel,
-    data_format='channels_last', kernel_initializer='lecun_uniform',
-    activation='relu'
+from DeepJet_models import binned3D_convolutional_classification_regression
+model = binned3D_convolutional_classification_regression(
+    inputs, output_shapes, config_args['conv_dropout']
 )
-
-#
-# Dense part
-#
-
-binned_info = Reshape((
-      int(binned_info.shape[1]) * 
-      int(binned_info.shape[2]) * 
-      int(binned_info.shape[3]), #coma to make it a tuple
-      ))(binned_info) #Flatten()(binned_info) seems to fuck up the tensor output shape, so I reshape
-X = concatenate([glob, binned_info])
-
-X = make_layers(
-    Dense, [[int(X.shape[1])]]+[[100]]*7, 
-    config_args['conv_dropout'], X,
-    dropout_at_first=True, kernel_initializer='lecun_uniform',
-    activation='relu'
-)
-X = Dropout(config_args['conv_dropout'])(X)
-
-# regression
-pt_sigma = Dense(2, activation='linear', kernel_initializer='lecun_uniform')(X)
-# classification
-flavour = Dense(output_shapes[0], activation='softmax',kernel_initializer='lecun_uniform')(X)
-model = Model(inputs=list(inputs), outputs=[flavour, pt_sigma])
-
 
 ####################################################
 #                                                  #
@@ -235,12 +141,31 @@ nvalepoch   = testd.getSamplesPerEpoch()
 testd.isTrain  = False
 traind.isTrain = True
 
+####################################################
+#                                                  #
+#                Warm start model                  #
+#                                                  #
+####################################################
+
+if args.warm:
+    print 'warm start from model directory %s' % args.warm
+    from keras.models import load_model
+    #model  = load_model('%s/KERAS_check_last_model.h5' % args.warm)
+    model.load_weights('%s/KERAS_check_last_model.h5' % args.warm)
+    shutil.copyfile('%s/trainsamples.dc' % args.warm, '%s/trainsamples.dc' % args.inputDir)
+    shutil.copyfile('%s/valsamples.dc' % args.warm, '%s/valsamples.dc' % args.inputDir)
+    traind = DataCollection('%s/trainsamples.dc' % args.inputDir)
+    testd  = DataCollection('%s/valsamples.dc'   % args.inputDir)
+
+
 print 'split to %d train samples and %d test samples' % (ntrainepoch, nvalepoch)
 #for bookkeeping
+traind.setBatchSize(config_args['batchsize'])
 traind.writeToFile(outputDir+'trainsamples.dc')
 testd.writeToFile( outputDir+'valsamples.dc')
 
 print 'training'
+
 
 ####################################################
 #                                                  #
@@ -256,7 +181,7 @@ model.fit_generator(
    validation_data = testd.generator(),
    validation_steps = testd.getNBatchesPerEpoch(),
    max_q_size = config_args['maxqsize'], #maximum size for the generator queue
-   class_weight = ['auto', 'auto']
+   #class_weight = ['auto', 'auto']
    )
 
 #Loss for classification should be < 0.48 regression loss < 2.8
